TracRaw_lvlock: store as int8_t to dodge a ROOT 6.38 RNTuple read bug
Symptom: ROOT::RNTupleView<std::vector<std::int32_t>>::operator()(i)
returns a vector whose .size() is correct but whose tail elements come
back as INT32_MIN / INT32_MIN+1 garbage for some events. The same file
read via uproot produces the correct values, and other int32 vector
fields (TracRaw_detectorsUsed, GenPart_pdgId, ...) read cleanly — the
bug appears to be specific to the small-range encoding RNTuple selects
for the lvlock field (~all 0/1 values).

Reproduced in a minimal C++ test against ROOT 6.38.04 on the key4hep
2026-04-08 stack, both via View and via the LoadEntry/REntry path.

Switching the writer schema to std::int8_t avoids the affected code
path entirely; values fit comfortably (LVLOCK is a few-bit flag). The
converter (delphi-improved-reco@feature/edm4hep-pipeline) reads as int8
and widens to int32 in Track_lvlock unchanged for downstream.

Saturating cast at the push site clamps values > 127 (none in practice)
to keep the int8 store well-defined under any future SKELANA change.

Verified end-to-end:
  - fresh 30-event nanoaod has TracRaw_lvlock typed std::vector<int8>;
    every entry reads correctly in C++ (no INT32_MIN garbage, total
    out-of-range count = 0)
  - converted EDM4hep file has 914 PFOs with 668 lvlock==0 + 246 >0,
    no -1 (clamp-from-garbage) entries.

+++ delphi-raw-nanoaod/include/raw_nanoaod_writer.hpp
@@ -233,7 +233,11 @@ private:
     std::shared_ptr<std::vector<std::int16_t>>            TracRaw_ndfVD_;         // Q(LMAIN+27)
     std::shared_ptr<std::vector<float>>                   TracRaw_chi2VDHits_;    // Q(LMAIN+18)
     std::shared_ptr<std::vector<std::int8_t>>             TracRaw_charge_;        // sign of Q(LMAIN+8)
-    std::shared_ptr<std::vector<std::int32_t>>            TracRaw_lvlock_;        // sk::LVLOCK at the matching VECP entry; 0 = passes IFLSTR=11/IFLCUT=3
+    // int8 (not int32) avoids a ROOT 6.38 RNTuple read-decoding bug specific
+    // to small-range vector<int32_t> fields (the C++ reader returns INT32_MIN
+    // garbage past element ~25 in some events; uproot reads the same file
+    // correctly). LVLOCK fits easily in 8 bits (values are 0/1/few-bit-flag).
+    std::shared_ptr<std::vector<std::int8_t>>             TracRaw_lvlock_;        // sk::LVLOCK at the matching VECP entry; 0 = passes IFLSTR=11/IFLCUT=3
     std::shared_ptr<std::vector<float>>                   TracRaw_vecpPx_;        // sk::VECP(1, vecp_i) — SKELANA-stored P_x for legacy-bit-exact parity
     std::shared_ptr<std::vector<float>>                   TracRaw_vecpPy_;        // sk::VECP(2, vecp_i)
     std::shared_ptr<std::vector<float>>                   TracRaw_vecpPz_;        // sk::VECP(3, vecp_i)
